requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
browser-use==0.5.7
bubus==1.4.7
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
fastapi==0.116.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.28.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.12.3
mdurl==0.1.2
minio==7.2.16
narwhals==2.0.1
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.1
openai==1.98.0
packaging==25.0
pandas==2.3.1
patchright==1.52.5
pillow==11.3.0
playwright==1.54.0
portalocker==2.10.1
posthog==6.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdf==5.9.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.2
streamlit==1.47.1
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uuid7==0.1.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
